Pipeline of uncertainty guidance diffusion model
The UGDM is composed of two steps: Pre-training and Two-stage guidance sampling. We train a classifier on a professionally annotated dataset. The pre-trained classifier is in charge of providing the pre-trained diffusion model with classifier and uncertainty guidance during the sampling process. The classifier guidance is a probability gradient belonging to target class, and the uncertainty is a uncertainty gradient based on the classification probability.